+++ src/asset/animation.cpp
