+++ src/asset/font.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+class Font {
+    // Wrapper for font
+};
