+++ src/asset/music.cpp
@@ -1,0 +1,1 @@
+#include "music.hpp"
