+++ src/asset/sound.cpp
@@ -1,0 +1,1 @@
+#include "sound.hpp"
