+++ src/asset/sound.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+class Sound {
+
+};
