+++ src/asset/texture.cpp
@@ -1,0 +1,1 @@
+#include "texture.hpp"
