+++ src/asset/texture.hpp
@@ -1,0 +1,5 @@
+#pragma once
+
+class Texture {
+
+};
